msmarco-v2-vector: better queries dataset by generating questions from Cohere API for each sample
