+++ cohere_v3_vector/operations/default.json
@@ -14,14 +14,14 @@
 {
   "name": "initial-documents-indexing",
   "operation-type": "bulk",
-  "corpora": "cohere-initial-indexing",
+  "corpora": "cohere-v3-initial-indexing",
   "bulk-size": {{initial_indexing_bulk_size | default(500)}},
   "ingest-percentage": {{initial_indexing_ingest_percentage | default(100)}}
 },
 {
   "name": "parallel-documents-indexing",
   "operation-type": "bulk",
-  "corpora": "cohere-parallel-indexing",
+  "corpora": "cohere-v3-parallel-indexing",
   "bulk-size": {{parallel_indexing_bulk_size | default(500)}},
   "ingest-percentage": {{parallel_indexing_ingest_percentage | default(100)}}
 },
