+++ cohere_v3_vector/track.json
@@ -11,7 +11,7 @@
   ],
   "corpora": [
     {
-      "name": "cohere-initial-indexing",
+      "name": "cohere-v3-initial-indexing",
       "base-url": "https://rally-tracks.elastic.co/cohere-msmarco-v2-embed-english-v3",
       "documents": [
         {
@@ -293,7 +293,7 @@
       ]
     },
     {
-      "name": "cohere-parallel-indexing",
+      "name": "cohere-v3-parallel-indexing",
       "base-url": "https://rally-tracks.elastic.co/cohere-msmarco-v2-embed-english-v3",
       "documents": [
         {
