+++ cohere_vector/track.json
@@ -11,7 +11,7 @@
   ],
   "corpora": [
     {
-      "name": "cohere-v3-initial-indexing",
+      "name": "cohere-initial-indexing",
       "base-url": "https://rally-tracks.elastic.co/cohere-miracl-en-corpus-22-12",
       "documents": [
         {
@@ -77,7 +77,7 @@
       ]
     },
     {
-      "name": "cohere-v3-parallel-indexing",
+      "name": "cohere-parallel-indexing",
       "base-url": "https://rally-tracks.elastic.co/cohere-miracl-en-corpus-22-12",
       "documents": [
         {
